refactor(dot-card-field): Migrate dot-card-field-label to dedicated component structure
