+++ core-web/libs/edit-content/src/lib/fields/dot-card-field/components/dot-card-field-label/dot-card-field-label.component.html
@@ -1,0 +1,11 @@
+<div class="flex gap-1 align-items-center">
+    <label
+        [attr.data-testid]="'label-' + $variableName()"
+        [class.p-label-input-required]="$isRequired()"
+        [for]="$variableName()">
+        <ng-content />
+    </label>
+    @if ($hint()) {
+        <i class="pi pi-info-circle text-xs" [pTooltip]="$hint()"></i>
+    }
+</div>
