+++ core-web/libs/edit-content/src/lib/fields/dot-card-field/components/dot-card-field-label/dot-card-field-label.component.scss
@@ -1,0 +1,20 @@
+@use "variables" as *;
+::ng-deep {
+    .p-tooltip .p-tooltip-text {
+        background: $foreground-active;
+        color: $white;
+    }
+
+    .p-tooltip.p-tooltip-right .p-tooltip-arrow {
+        border-right-color: $foreground-active;
+    }
+    .p-tooltip.p-tooltip-left .p-tooltip-arrow {
+        border-left-color: $color-palette-black-op-50;
+    }
+    .p-tooltip.p-tooltip-top .p-tooltip-arrow {
+        border-top-color: $color-palette-black-op-50;
+    }
+    .p-tooltip.p-tooltip-bottom .p-tooltip-arrow {
+        border-bottom-color: $color-palette-black-op-50;
+    }
+}
