+++ core-web/libs/edit-content/src/lib/fields/dot-edit-content-binary-field/components/dot-binary-field-wrapper/dot-binary-field-wrapper.component.ts
@@ -6,7 +6,7 @@ import { DotMessagePipe } from '@dotcms/ui';
 
 import { DotCardFieldContentComponent } from '../../../dot-card-field/components/dot-card-field-content.component';
 import { DotCardFieldFooterComponent } from '../../../dot-card-field/components/dot-card-field-footer.component';
-import { DotCardFieldLabelComponent } from '../../../dot-card-field/components/dot-card-field-label.component';
+import { DotCardFieldLabelComponent } from '../../../dot-card-field/components/dot-card-field-label/dot-card-field-label.component';
 import { DotCardFieldComponent } from '../../../dot-card-field/dot-card-field.component';
 import { BaseWrapperField } from '../../../shared/base-wrapper-field';
 import { DotEditContentBinaryFieldComponent } from '../../dot-edit-content-binary-field.component';
